web/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
httpx==0.27.0
jinja2==3.1.4
python-multipart==0.0.9
itsdangerous==2.2.0
pyyaml==6.0.1
